Add `paperless` user
One can map the UID and GID of the created user with the environment
variables `USERMAP_UID` and `USERMAP_GID`.
The container will then on start change the IDs accordingly, and `chown`
the `/usr/src/paperless` directory as well as the `$PAPERLESS_CONSUME`
directory.

Be aware that if you mount any of those directories in your local
filesystem, permissions will be changed!